The function to count the number of letters
Function counts how many times a letter matches in text regardless of case.
Two solutions. In the first one through two if, where lower case is compared first, then upper case.
In the second solution, the letters are first reduced to the same case, and then matches are already considered.