Function with while, with n-th signs in upcase
I finally completed this quest from Hexlet. The function must return text in which the characters specified by the second parameter are converted to uppercase. The whole difficulty was that in the condition of the logical operator I divided the string by a number, because of this, JS counted the characters incorrectly.