Content basics Practic 9
So the ninth css practice, pseudo-classes, helps us to style some element state. For example, hovering over an element, focusing on an element, clicking on an element. The main properties that help us select multiple elements at once and style them: :nth-child(even/odd), :nth-of-type. For an exercise on this topic, the task is ideal: draw a chessboard.