Positioning in css Practice 1
We are starting a course on positioning in css, today, according to the assignment, it was necessary to add a property that moves the card when hovering over it. This positioning is called relative positioning. in css this rule is called position: relative and is controlled by top, right, bottom, left.