Positioning in css Practice 3
So, the third practice of positioning in line, the whole trick is that inside an element with relative positioning, we put an element with absolute positioning, for example, a cross in a pop-up window. so for example, we can stretch the entire block relative to the parent block using the same: top, right, left, bottom with a value of 0.