Positioning in css Practice 6
Coming to the end of the course on positioning elements in CSS, this lesson on the overlay of elements. The main property is the z-index, that is, the Z-axis, the higher its value, the higher the element is. This property only works on elements that have a position. Opacity properties were used to make the task transparent.