JS Arrays Practic 9
The topic of control structures, in this lesson we talked about two: Break, Continue.
Break - exits the loop.
Continue - allows you to skip the iteration of the loop.

The break and continue are intended to increase flexibility in managing the traversal process. In practice, you can always build code without them - most likely, it will be even easier. Avoid these constructs whenever possible.