JS Arrays Practic 12
The task was to check the text for the presence of certain words, if there were any, then they are replaced by another word. The following methods were applied: split(), join(), includes().
First, we create an array of all words from the text, then we create an empty array in which the answer will be, now we write a for..of loop, in which we create a constant and write a ternary with it for the presence of stop words in the array, if there is a word, we change it to symbols , if not, we leave it as it is, push the result into our empty array, return the result to the function and add spaces with join().