JS Arrays Practic 17
The task was to find intersections in two arrays. My solution is very simple, create an empty array, write a loop in a loop and check each element, write intersections to an empty array so that there are no duplicates, check whether there is such an element in the array, skip iteration.