JS Arrays Practic 19
A simple task to use destructuring and the Rest operator, which splits our array or string into several. The task was to find the maximum element in the array, cut our argument into the first element and the remaining values ​​in the array, write the first element to a variable, then compare it in a loop with the rest of the elements, if the element is larger, then assign a new value to the variable.