Command line basics 3 practice
Practice 3, dedicated to the talking cow, according to the task, we already had to apply the argument so that instead of the cow there was a tux penguin that says "I love Hexlet". To navigate through directories, the command cd ../, ls option show hidden files -a, -l show additional information on files.