Command line basics 10 practice
The tenth practice in terms of environment variables, you can see the set variables using the env command, variables are very convenient, for example, with their help, the cd command understands where the home directory is located, their values ​​​​can be changed in the same way as in programming languages, the syntax is the same. You can view the value of a variable with the echo $NAMEVAR command. If you write an export, you can change the values ​​of variables and create new variables separated by a space.