Command line basics 11 practice
Command line basics, 11 practice, history of commands, the entire history of commands can be found in the .bash_history file, the number of commands that will be stored in this folder can be determined in the HISTFILESIZE variable. You can also view the history more simply by writing the history command, if you add a number to the command it will show the same number of recent commands. There is also a recursive search: the command Ctrl + r and start is included in the beginning of the command.