Command line basics 12 practice
Users and groups. Practice 12, using the "whoami" command you can find out the username, the "ps" command displays a report on running processes, the "ps aux" command shows which process is running under which user. There are also several ways to see your ID, through the "id" command and through "cat /etc/passwd".

The task is to display only the id of a specific user.