Command line basics 14 practice
The topic of access rights, There are three types of access, read (r), write (w), execute (x), files and directories can be assigned access, you can do this through the chmod command, there are different ways, just my solution with the teacher's is slightly different in syntax. According to the assignment, it was necessary to reassign access to the file for different users (owner (u), group members (g), others (o)).