Command line basics 15 practice
The final practice on the Basics of the command line, the topic "Execution of programs", in order to find out where the executable file is located, you can use such commands (type[namefile], whereis[namefile], which[namefile]). To execute a file, we use the following syntax: ./[name of the executable file]. According to the task, it was necessary to change the access rights to the file, add "execution" to it, then run it. In order not to copy the result from the command line, you can use stream redirection to the file where you want to copy the result. (./program > solution).