Introduction to Git 1
Let's start the next course, this time it's Git, in short, we need git for collaborative and secure development, in the first practice we needed to add several files with text to the repository and commit them. The commands I used to get the job done are: cd (to move between directories), echo (to create files), (>) to direct the flow, git status (to check for uncommitted files), git add (to add a file to the index), git commit -m (commit the file).