Introduction to Git 2
Integration with GitHub. A very rich second practice, in addition to the main task, in parallel you set up a git on your computer. So, in the first part of the lesson, we created an empty repository on GitHub, the repository will contain commands for connecting it to the repository on the computer, before doing this, do not forget about initialization through (git init), then we enter all the commands, the last push the changes (git push). We now have a copy of our GitHub repository. The repository can be cloned: Copy the path to it in GitHub and write it with the "git clone" command. There is also a command that pulls all the changes from GitHub: git pull --rebase.
The task was to make changes to the files and push them to GitHub.