Introduction to Git 4
In this lesson, we analyze changes, this is done through the command: git diff and git diff -staged for files that are already in the index, it is convenient to track what has changed in the file. The task was to rename the first file, and delete the second and commit them separately with an analysis of the changes.