Introduction to Git 7
the seventh practice made my task a little more difficult, since I needed to restore the old commit, but the problem was that after deleting I did git init again and my commit history was erased, is it possible to restore the commits somehow differently, I still don’t know . The topic of the lesson is simple: Reverting commits. The command git revert [commits' id hash] - in fact, it creates a new commit, the opposite of the reverted one, very convenient.
The git reset command removes the newly added commit, and this option leaves no trace in the history, it is important to do this before pushing to GitHub, that is, before git push, as this will cause problems.
To completely remove a commit, you need to register the --hard flag, if we want to remove several commits from the end, then the second flag is HEAD ~ (the number of commits to be deleted).