Introduction to Git 9
The index topic is a staging area that contains modified files on the path from the working directory to the repository. If we want to commit a file right away, we can write it like this: git commit [namefile] -m 'comment'. If I want to add all changes to the index, then git add . . If we need to commit some specific parts, then this is done through interactive indexing.