Trial "Conflicts"
Cool task, on conflicts when merging branches. The main course did not have this topic, so the merge turned out to be new material for me. The commands I used in the task are: (move to another branch) git checkout <filename>, (merge branch with the branch you're currently on) git merge <branchname>. If there are conflicts when merging branches, then these files appear in the git status and mark the places with markers that we need to fix, then add everything to the index and commit.