Trial "Calendar"
The task, as always, was to lay out the layout, this time the calendar, nothing complicated, but a little inattention cost me several hours of training. Since this is just the basics of content layout, there are no flex, grids, etc., so the properties used are the same as in the previous tests.