JS: Objects Lesson 1
Today I'm starting to study objects in JS, this is the most fancy data type in JS, the theory tells about the syntax and there are already interesting nuances, you can access the elements of an object in several ways: object name. property && object name ['object property'] - this way access allows you to set the name dynamically, it will not work through a dot.