JS: Objects Lesson 2
The second lesson on objects, topic Modification, in order to create or update property values ​​in an object, one operation is used - assignment, if the property does not exist, then it will be added to the object, if it exists, it will be overwritten. You can delete an element from an object using the delete operator (delete name object. object property). However, removing properties is bad practice.