JS: Objects Lesson 4
The fourth practice, Objects in Action, is to implement a function that takes the site name as input and returns information about it. For the task, I used the lodash library and the following properties: _head(return the first element of the array), _last(return the last element of the array), .split('something')(return the string before 'something'.