JS: Objects Lesson 5
This is the fourth practice in Objects, and the topic is "Checking for the existence of a property", to check we need the Object.hasOwn() method, which returns us true or false, and a little more about the merge null operator. The challenge is to implement a function that takes a string of words and returns an object. Properties are words, and values ​​are the number of repetitions of those words in a string.