JS: Objects Lesson 6
This is the sixth practice in the course Objects in JS, topic: Traversing object properties. The Object.keys() method allows you to get an array of all the keys of an object (to make it clearer "properties"), if we need not properties, but values, then they can be obtained using Object.values(), there is also Object.entries() this the method will return us an array with arrays containing properties and their values. And then with the help of destructuring, you can access your favorite data. I also learned from this lesson that objects do not have a strict order, like arrays, there is an order, but it is of a different kind.