JS: Objects Lesson 7
Hooray, I managed to defeat this lesson on the topic "Nested Objects", in the task it was necessary to implement a function that returns a value from a chain of arrays, if it exists. That is, we created a cycle from an array with a chain and in it, through the Object.hasOwn() method, we checked whether our object has such a property, if so, then we continue to fall into the depth of the object further, if there is no such property, then we return null. The lodash library has a get() function for this, which will solve our problem in one line (_.get(obj, ['one', 'two', 'three'], 'defaultValue');), but the task was to implement this algorithm alone. Also, in practice, they talked about the Optional Sequence Operator (provides the ability to receive any data from an object of any depth) "obj?.one?.two?.three"