JS: Objects Lesson 8
The eighth practice "Merge". To fill in the data of one array into another array, the Object.assign(obj1, obj2) method is used, everything is simple, but there are nuances, it performs only a superficial merge. Nested objects are not compared, but simply replaced. We can also use _.pick(obj, [something]) to select specific keys that we want to replace in our object.