JS: Objects Lesson 9
The ninth lesson on objects, Cloning. You can clone an object using the Object.assign() method (fill an empty object) or using the clone() function of the lodash library, there is a nuance that these methods do not affect nested objects, they end up in a new object by reference from the old one, such cloning called surface. If we want to do a deep clone, we can use lodash's cloneDeep() function. There is a serious disadvantage, if the object is very large, then this affects performance, so often deep cloning is not suitable.