JS: Functions Practic 1
JS: Functions, first practice "Pure Functions". A pure function is a function that has no side effects (no interactions with the environment or changes to global variables) and is deterministic (always produces a predictable result). Therefore, well-designed programs try to isolate side effects in a small part of the application so that most of the code is clean. A small example, demonstrated in practice: a function with side effect isolation.