JS: Functions Practic 2
Topic of the second practice: Packing arguments through the Rest operator. As far as we remember, rest returns an array with the data passed to it (that is, it packs them into a separate array), which is convenient in our case if we need a different number of arguments in the function each call. Regarding the syntax, remember that the ...rest operator is always written at the end of (x, y, ...rest).