JS: Functions Practic 3
So, the third practice on the topic of the spread operator (argument unpacking) does the opposite of the "remainder" operator. The "spread" operator spreads an array over its arguments. In the assignment, it was necessary to use it in a loop in order to solve the problem correctly. This is clearly shown in the teacher's decision.