JS: Functions Practic 4
The fourth practice on the topic "Objects of the first class", they denote elements that can be passed to functions, returned from functions and assigned to variables (or constants). The job is to implement a function that returns the reversed n- number of characters from the end of the string. The teacher's decision was made through cycles, and I did it through converting first the string to an array, and then the methods. The teacher also made a check inside the second function, for an empty string and its length, which is more correct.