JS: Functions Practic 5
So we continue to get acquainted with functions, today we begin to study higher-order functions. And for execution, we need to sort the array and filter the number of values ​​​​with the second parameter that will fall into the result. I was able to sort the array using the built-in sort() method, but it would be more correct to use the lodash library and the _.sortBy function, as in the teacher's solution.