JS: Functions Practic 6
We continue to get acquainted with higher-order functions, one of the more common ones: map () is a mapping, that is, we replace the description of the entire logic of bypassing the input data through cycles with one line, and we get the same result - a new collection. The assignment was to return a collection of objects with children, and a small flat() method helped us subarrays 1 level up. The teacher's decision seemed more organic to me, because the code looks more visual through destructuring.