JS: Functions Practic 8
Today I was struggling with another higher order function - reduce(), which is used for data aggregation. The reduce() function takes two parameters as input (the handler function and the initial value of the accumulator), the same accumulator is returned outside as the result of the entire operation. The function passed to reduce() is the most important part and the key to understanding how the whole aggregation mechanism works. It takes two values ​​as input. The first is the current value of the accumulator, the second is the current item being processed. The task of the function is to return the new value of the accumulator.