JS: Functions Practic 9
So, we have 9 practice in line and it is called
conditionally "Signals", because drag and drop of data goes along the chain, as in the design of electrical circuits. In fact, the association turned out to be very clear and understandable. This way of working with collections in JS is the main one, such code is much easier to analyze. According to the assignment, we get an array with email addresses, and from them we need to return the number of free domains. To do this, we needed three higher-order functions (map(), filter(), reduce()).
With map() we left only our domains from the collection without names, then with filter() we compared and left only free domains, and with reduce() we were able to count how many of each domain were in our collection.