JS: Functions Practic 11
And this is an extreme practice in the course: An iterative process, an iterative process itself, also describes recursion. As we remember, recursion does not perform any calculations until it reaches the base case, imagine how much memory it takes to calculate factorial (100), unlike recursion, the iterative process stores information about the calculations and if it stops at a certain moment, it will give a part of the result.