JS: automated testing 2
The second practice in automated testing is about assertions, the test checks the assertions, this is the key part, it checks the function of the code. Node.js has a built-in assert() module that replaces all comparisons with if , calling our tests in a more concise and understandable way. To work with assert, it must be imported. There is another difference, if we want to compare values ​​in objects, then we need to write an assertion through assert.deepEqual(), since the strictEqual(actual, expect) function checks for equality by reference.