JS: automated testing 5
Matchers, technically, this approach aims to make the tests look like a verbal description of the task being performed. This makes it possible to use them as documentation for people who do not know how to program. Matchers have replaced regular function statements in many languages. Their problem is that there are a lot of them and it is difficult to remember everything, so you have to constantly look at the documentation and remind.