JS: automated testing 6
We continue testing, topic: Unit tests. 1. First we test the main functionality, the first test should always check the positive scenario - the one in which the main functionality of the component under test is involved. If there is only one scenario, then there is no need to separate it into a separate test, since this will create code duplication. 2. We test additional functionality, no need to try to find all the options, learn to find the necessary minimum. 3. Edge cases, such as empty objects or arrays. If there is a bug in the code for which there was no test, first write a test that resolves the bug and then fix it. This is the only way to maintain a sufficient level of reliability without turning development into continuous error checking.