JS: automated testing 7
Subject: data preparation. As a rule, if we need to test several functions or a library, then we need a collection that is suitable for testing most or even all functions, it needs to be moved to the module level, outside the test functions. This solution eliminates unnecessary duplication. overgeneralization, leading to the complete elimination of duplication, introduces implicit dependencies into the code. Modifying this collection will almost certainly break most of the tests related to its structure, number of elements, and their values. The main conclusion from this: it is necessary to eliminate duplication. But it is important not to cross the line, after which generalization becomes more of a hindrance than a help.