JS: automated testing 8
In this lesson, we talk about good and bad testing practices. 1. Mutual influence of tests
One of the key rules: tests should not affect each other. This means that any test is performed as if other tests do not exist in nature. 2. Conditional constructions in the test. Any branching within tests is actually multiple tests within a single test. You need to get rid of this and never write like that. 3. Test out of tests. The task of beforeEach is to prepare the data and environment for testing, and the task of test is to call the code that is being tested and perform checks. 4. Too much detail. More often than not, the only result of this separation will be more code and more difficult refactoring in the future, when there is really a lot of code. 5. Deep nesting. 6. Code with tests takes longer to write than code without tests.
This is a very interesting question by which you can understand how well a programmer can write tests. Even though some types of testing are really complex and require additional time, daily tests that are written along with the code should lead to faster development.