JS: automated testing 9
Cool topic, Code coverage with tests. To form a pivot table and see code coverage by jest tests, you need to write the command in the line (npx jest --coverage), the coverage percentage will appear in the table, this does not mean that 100% gives us a guarantee that the code will work in all situations. Logical errors in code cannot be traced by coverage alone. 80% is pretty good coverage. According to the assignment, it was necessary to look at the coverage of 3 functions and add tests to the uncovered code.