JS: automated testing 10
My last testing practice. Topic: Development through testing. Testing before writing code has a powerful advantage. It forces the programmer to think first of all about how it will be used. And not about how beautifully it implements everything inside. Competent interfaces are the key to success. TDD, as conceived by the inventor of this technique, implies that the entire development consists of a repeating cycle, where at each iteration a test is written that does not pass, then code is added that satisfies this test. After that, everything repeats. So, step by step, the application is built.