Trial " Last word length"
Trial "The length of the last word". A little practice between lessons. The task is to find the length of the last word in the string, there are extra spaces at the end of the string, we remove them with the trim () method and cut our string into an array, take the extreme value through the index or the at () method and return its length.