JS: Abstraction with data 2
Abstraction with Data, Second Practice: Array Semantics. When solving this or that problem, one should remember that in addition to syntax there is another important thing called semantics. An indexed array is suitable for collections of the same type of data such as: List of stop words, list of users, list of course lessons, list of chess moves (order is important). And for an associative array, any single entity is suitable: User, rate, coordinate points, payment. Using an object immediately gives another extremely important advantage - extensibility. An indexed array used as a structure is fragile. You cannot swap the value of the arguments - all the code that counted on a certain order will break, or you will have to rewrite everything. Expanding it just won't work either: some of the code will, of course, continue to work, but some may break (for example, [x, y] = point). And the use of an object does not rely on the order of the keys, and certainly does not depend on their number. You can add a new key at any time, and the program will almost certainly remain functional.