JS: Abstraction with data 3
We continue to talk about abstraction. The essence of abstraction is that we hide the internal implementation. We can say that creating an abstraction with data leads to hiding this data from external code. This gives us the ability to change the internal implementation of the data without rewriting the code that uses this data and the code is easier to read at the expense of function names (provided they are well named). In practice, we need to implement several functions using abstraction.