JS: Abstraction with data 4
We continue to practice with abstractions, today I more or less figured out what Interfaces are - in simple words, this is what the user interacts with the abstraction through. For the assignment, I needed to write several functions that return us points converted from a polar coordinate system to a polar one.