Trial "Chunking"
A simple task on the slice () method, two parameters are given, the first is an array, the second is the number by which the entire array must be divided. We write a for () loop with a step on the second parameter and push our array along with the slice, after which we return a new array.