JS: Abstraction with data 5
When developing complex systems, an approach is used - level design. It lies in the fact that the system is given structure with the help of successive levels. Each of the levels is built by combining parts that are considered elementary at this level. Parts that are built at each level work like elementary parts at the next level. In practice, I needed to write a number of functions, the first one creates a rectangle, then three selector functions for convenient data access, and the fifth function checks if the center of the coordinate axis is in the rectangle.