JS: Trees 1
The key feature of a tree structure is that it is recursive. A tree consists of nodes and edges between them. Edges do not exist in reality, they are only needed to visualize the connection and, if necessary, describe it. Nodes are divided into two types: internal (those that have children) and leaf nodes (those that have no children). In the case of a file system, leaf nodes are represented by files, while internal nodes are directories. In addition, trees distinguish the concept of depth (depth), which determines how many steps you need to go through the vertices from the root in order to reach the current one (the one we are looking at). The most flexible and convenient view of trees is objects. In such a rustic place it's an object, arrays are only used to store a list of children.