JS: Objects Lesson 12
Hash tables, in this final lesson we will talk about what happens under the hood of objects, we will analyze how they are arranged. Associative arrays or the most popular objects in application programming. in addition to the key-value, in our usual representation, a hash is first formed into objects, which is the very link by which you can get the value through the key. This very hash is formed by itself, we do not see it in the object, but it is:). It should also be mentioned that there are collisions and many ways to fix them, but the essence of them is that a hash is a limited number of characters, and the number of keys is infinite, so there are cases when the keys are different, but the hash is the same, this is a collision. Our task is to figure out how the process inside objects works, we need to write a number of functions, creating an empty array, creating a hash, checking for collisions and basic functions: setting a value to our array by key and returning a value by key.