Create Trial "Matrix mirroring"
The task is to mirror elements in an array. I created two loops to iterate through all the elements and truncate them to the middle. Copied the element using the spread operator, reversed it with reverse(), then added it to a new array, and finally used the flat() method to raise the array one level.