JS: Trees 2
With the help of trees, we can display a virtual file system. By nesting calls to mkdir and mkfile within other mkdir, any file structure can be obtained. By assignment, we need to implement a small file system. Next, we will implement everyday operations for working with it: counting free space, searching for files and directories, and others.