JS: Trees 3
We continue to learn how to interact with trees using the example of a virtual file system. Changing the name of a file, sorting the contents of a directory, updating the contents of a directory, deleting files, all this is done by making a copy of the object with our changes. The job is to halve the size property if the filename contains .jpg and then return the same object.