JS: Trees 4
Tree traversal is a step-by-step traversal of all tree elements. There are several tree traversals, in this course I study depth traversal. The point of this traversal is to go as deep as one subtree as possible. This algorithm naturally falls on the recursive solution and turns out by itself. In the assignment, we need to lowercase all the files in the object. First, let's save our names and meta data into separate constants. Then we check if our element is a file, then we return it with the toLowerCase() method. Then we copy all the elements into a separate constant. We create a constant in which we launch a recursive traversal of our elements, as a result we collect names, new elements, and new meta data in one directory, done))