JS: Trees 5
Data aggregation is the most important operation when working with trees. Count the total number of files in a directory, the total size of all files, get a list of all files, find all files by a pattern - these are all examples of data aggregation. The task is to count all files that have "." at the beginning of the name. Then we create a constant for recursion to check the entire tree. Now we create a constant with the launch of recursion, if our constant is an object. And now we return the amount.