JS: Trees 6
Another lesson on aggregation. Now, according to the assignment, we must return the list of nested nodes in the specified directory one level, as well as the space they occupy. If not, then we create a constant with the children of our node and run our test function recursively, finally summing up all returned size properties. Next, we write the second function, in it we create a constant for the children of our object, then we create a constant in which the result will be stored, in it we run .map with each child, which returns the name and result of the first function. At the end, we return the constant result with the sort method in reverse order.