JS: Trees 7
Let's talk about accumulation. There are situations when we need information that we can get when traversing a tree. For example, refers to the full path to the file or the depth of the current node. Then an additional parameter can help us, which will collect all the information for us. According to the assignment, we need to implement a function that takes a tree and a substring, we need to return the paths to the files, in the name of which the substring is contained. We write our main function. Inside, we create a check recursive function with parameters: the node to be checked and the accumulator. Create two constants name and children. Then our check, if the child is a file, then we check it for the presence of a substring, if it is, then we add it to the accumulator, if it is not, then we return an empty array. Now we create a constant - a new accumulator, in which we write the name of the current directory. Our inner function is ready, close it and return in our main function a call to the inner function with the original tree and an empty accumulator.