JS: Introduction to OOP 1
We start the course on the basics of OOP, And the first thing we will analyze is Encapsulation. Encapsulation is the combination of functions and data within a single structure, the data of which is hidden from the outside world. Such functions are called methods. Encapsulation is both a union and a hiding where it exists. Where there is none, it is simply a union. Working with methods instead of functions leads to one unexpected effect - it becomes possible to implement autocompletion of methods in editors. In JavaScript, regular functions and methods fit right in together. Now let's analyze the task, we need array with the intersection of "friends" of two users. Therefore, first, using the getFriends() method, we get two constants with data about friends. Now with the help of .map(), we get the id data of the friends of one of the users. And finally, we return .filter() with a different user.