JS: Introduction to OOP 2
Context (This). To fully study OOP in JavaScript, you need to understand such a concept as the this context. Nearly everything else builds on context, including how methods and classes work. Through a special "this" keyword called context. Within methods, it refers to the current object that the method is bound to. The key difference between this in JavaScript and this in other languages. In JavaScript, a method's this can change. Calling the same function from another object will change the object referenced by this. This feature is called late binding. The this value refers to the object from which the method is called. The teacher's solution contains an abbreviated syntax.