JS: Introduction to OOP 3
When a function is passed to another function, it loses its connection to the object itself and its this no longer points to the object. In order for our object not to be lost, the bind () method is used, its task is to bind the function to some context. The result of executing bind() will be a new function that works like the original function, but with a context bound to it. Apply & Call bind() is useful where the context binding and the function call happen in different places and usually at different times. Sometimes the call to functions that use this internally occurs immediately along with the context binding. You can use the apply() and call() functions specially created for this: These functions do two things internally: they change the context and immediately make a function call. The only difference is how they work with the arguments of these functions: apply() takes arguments as an array as the second parameter, and call() waits for positional arguments as input. In this exercise, you need to replace one function with another to get functionality similar to bind().