JS: Introduction to OOP 4
Now a few words about how "this" differs in an arrow function from a regular one. The context of regular functions depends on where they are called, while the context of arrow functions depends on where they were defined. An ordinary function has, as expected, associated with the context of the object on which it is called. But in the case of the arrow function, this did not happen. Why? An arrow function does not have its own context, it is associated with the lexical environment, that is, the function within which the arrow function is defined. This is a very important point. It is the top-level function that sets the context of the arrow function, and nothing else. And this behavior cannot be changed using the call or bind functions.