JS: Functional Programming 1
In parallel with the OOP course, I am taking a course in Functional Programming. In the first lesson we talk about cyclomatic complexity. The cyclomatic complexity of a piece of program code is the number of linearly independent routes through the program code. If the source code does not contain any branch points or loops, then the complexity is one, since there is only one route through the code. The approach we'll learn, also called "patterning," helps structure a function better and sometimes reduce cyclomatic complexity. Refactoring should reduce the level of nesting, which reduces the likelihood of making mistakes in the logic. In practice, we get a string parameter and have to determine if it reads the same way backwards. We need to solve using recursion, first we write the condition, if the string is less than 2, then return true. Then we return the ternary operator, if the first character is not equal to the last return false, otherwise we run our function recursively passing a substring into it without the first and last character using the substring() method. Done.