JS: Trees 8
My last practice is in a course on trees. In this tutorial, we'll get to know the HTML markup tree. To make it convenient for us to work with this structure, we first describe properties for each tag. Then we define the type of nodes. Now we have the opportunity to work with our structure. We can filter empty tags or for example change values ​​in properties. My task was just to change the values ​​of the "class" property to a new one. So I'm writing a function with three parameters: our tree, the old class, and the new class. I do a deep copy so that our object does not mutate. Now I run .map on each node and do two checks: first, if the node has a class, then we replace its value with a new one. And secondly, if our tag can have children, then we recursively run our function to this node in order to traverse the entire tree. And at the end we return our new tree with the spread operator and the children that passed the checks.