JS: Introduction to OOP 5
As we learned earlier, if we are dealing with entities, then we just need abstractions, so we hide all the code structure that we do not need and leave only the names that we operate in the future, which is quite convenient. But in order not to create unnecessary methods every time using abstraction, we can remove these functions from the constructor and take them out. So that we can later use our abstraction to create new objects, we use a "new" operator.