JS: Introduction to OOP 6
Prototypes are the mechanism that has a major impact on how JavaScript objects work. In JavaScript, every object has an associated prototype. A prototype is a regular object stored in a special [[prototype]] service field (this field cannot be accessed directly). It can be extracted. The prototype can be accessed not only from objects, but also from the prototype property of a constructor. A prototype is an object found in the prototype property of a constructor function, not the constructor itself. It is quite easy to check the work of prototypes by changing them. Prototyping properties is the right way to create custom abstractions in JavaScript. Any new abstraction that we need in code should look like a constructor and a prototype filled with the right properties. Prototypes are an incredibly flexible mechanism that allows you to change anything from anywhere in the program. For example, we can add or replace any methods in any objects of the language itself. But if you use this mechanism wisely, you can get a lot of good. For example, prototypes are widely used in testing, they help to replace calls to unwanted methods that make HTTP requests. In practice, we need to Implement the "Money" abstraction. She knows about the currency of money, allows you to convert them into other currencies, perform arithmetic operations and format the output. Methods that we need to implement: create a money object (the Money() constructor function), return the amount (getValue()) of money, return the currency (getCurrency()), convert, sum money in different currencies (exchangeTo(currency)), return money in a convenient form (format()).